server/requirements.txt
Flask>=2.2.0
Flask-SQLAlchemy
psycopg2-binary==2.9.5
python-dotenv==0.19.0
Flask-CORS==3.0.10
pytz==2025.1
six==1.17.0
tzdata==2025.1
bcrypt
